requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
python-multipart==0.0.12
qrcode[pil]==7.4.2
Pillow==10.4.0
aiosqlite==0.20.0
